Clean up code changes table
